+++ src/polymer/content-to-slot.ts
@@ -13,6 +13,7 @@
  */
 
 import * as dom5 from 'dom5';
+import * as parse5 from 'parse5';
 import {treeAdapters} from 'parse5';
 import {Document, ParsedHtmlDocument, Severity} from 'polymer-analyzer';
 
@@ -93,6 +94,19 @@ class ContentToSlot extends HtmlRule {
   async checkDocument(parsedDocument: ParsedHtmlDocument, document: Document) {
     const warnings: FixableWarning[] = [];
 
+    // TODO(rictic): rather than just hard coding translation rules as above,
+    //    we should also support converting based on the old-content-selector
+    //    attribute.
+    this.convertUsages(parsedDocument, document, warnings);
+
+    this.convertDeclarations(parsedDocument, document, warnings);
+
+    return warnings;
+  }
+
+  convertUsages(
+      parsedDocument: ParsedHtmlDocument, document: Document,
+      warnings: FixableWarning[]) {
     const references = document.getFeatures({kind: 'element-reference'});
     for (const reference of references) {
       const contentDescriptors = config.get(reference.tagName);
@@ -150,7 +164,10 @@ class ContentToSlot extends HtmlRule {
       if (fix.length > 0) {
         const warning = new FixableWarning({
           code: 'content-to-slot-usage-site',
-          message: ``, parsedDocument,
+          message: `Deprecated <content>-based distribution into ` +
+              `<${reference.tagName}>. ` +
+              `Must use the \`slot\` attribute for named distribution."`,
+          parsedDocument,
           severity: Severity.WARNING,
           sourceRange: parsedDocument.sourceRangeForStartTag(reference.astNode)!
         });
@@ -158,15 +175,67 @@ class ContentToSlot extends HtmlRule {
         warnings.push(warning);
       }
     }
+  }
 
-    // TODO(rictic): rather than just hard coding translation rules as above,
-    //    we should also support declaring the legacy content selectors
-    //    on a <slot> element.
-
-    // TODO(rictic): convert <content> *declarations* to <slot>, in addition
-    //     to just updating *usage* of such elements.
-
-    return warnings;
+  convertDeclarations(
+      parsedDocument: ParsedHtmlDocument, document: Document,
+      warnings: FixableWarning[]) {
+    for (const element of document.getFeatures({kind: 'polymer-element'})) {
+      const domModule = element.domModule;
+      if (!domModule) {
+        continue;
+      }
+      const template = dom5.query(domModule, p.hasTagName('template'));
+      if (!template) {
+        continue;
+      }
+      const contentNodes = dom5.queryAll(
+          treeAdapters.default.getTemplateContent(template),
+          p.hasTagName('content'));
+      const slots = new Set<string>();
+      for (const contentNode of contentNodes) {
+        const warning = new FixableWarning({
+          code: 'content-to-slot-declaration',
+          message:
+              `<content> tags are part of the deprecated Shadow Dom v0 API. ` +
+              `Replace with a <slot> tag.`,
+          parsedDocument,
+          severity: Severity.WARNING,
+          sourceRange: parsedDocument.sourceRangeForStartTag(contentNode)!
+        });
+        const attrs = [...contentNode.attrs];
+        const selectorAttr = attrs.find((a) => a.name === 'select');
+        const selector = selectorAttr && selectorAttr.value;
+        let slotName = null;
+        if (selector) {
+          slotName = slotNameForSelector(selector);
+          while (slots.has(slotName)) {
+            slotName += '-unique-suffix';
+          }
+          slots.add(slotName);
+          attrs.unshift({name: 'name', value: slotName});
+          attrs.push({name: 'old-content-selector', value: selector});
+        }
+        const slotElement = treeAdapters.default.createElement('slot', '', []);
+        for (const {name, value} of attrs) {
+          dom5.setAttribute(slotElement, name, value);
+        }
+        dom5.removeAttribute(slotElement, 'select');
+        const fragment = parse5.treeAdapters.default.createDocumentFragment();
+        dom5.append(fragment, slotElement);
+        warning.fix = [
+          {
+            replacementText: parse5.serialize(fragment).slice(0, -7),
+            range: parsedDocument.sourceRangeForStartTag(contentNode)!
+          },
+          {
+            replacementText: '</slot>',
+            range: parsedDocument.sourceRangeForEndTag(contentNode)!
+          }
+        ];
+        warnings.push(warning);
+      }
+    }
   }
 }
 
@@ -188,6 +257,14 @@ function simpleSelectorToPredicate(simpleSelector: string): dom5.Predicate {
     return () => true;
   }
   return p.hasTagName(simpleSelector);
+}
+
+function slotNameForSelector(selector: string) {
+  const identifierMatch = selector.match(/[a-zA-Z-_0-9]+/);
+  if (identifierMatch) {
+    return identifierMatch[0];
+  }
+  return 'rename-me';
 }
 
 registry.register(new ContentToSlot());